+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/client/ByProjectKeyAsAssociateByAssociateIdInBusinessUnitKeyByBusinessUnitKeyShoppingListsByIDRequestBuilder.java
@@ -1,0 +1,65 @@
+
+package com.commercetools.api.client;
+
+import java.util.function.UnaryOperator;
+
+import io.vrap.rmf.base.client.ApiHttpClient;
+import io.vrap.rmf.base.client.utils.Generated;
+
+@Generated(value = "io.vrap.rmf.codegen.rendering.CoreCodeGenerator", comments = "https://github.com/commercetools/rmf-codegen")
+public class ByProjectKeyAsAssociateByAssociateIdInBusinessUnitKeyByBusinessUnitKeyShoppingListsByIDRequestBuilder {
+
+    private final ApiHttpClient apiHttpClient;
+    private final String projectKey;
+    private final String associateId;
+    private final String businessUnitKey;
+    private final String ID;
+
+    public ByProjectKeyAsAssociateByAssociateIdInBusinessUnitKeyByBusinessUnitKeyShoppingListsByIDRequestBuilder(
+            final ApiHttpClient apiHttpClient, final String projectKey, final String associateId,
+            final String businessUnitKey, final String ID) {
+        this.apiHttpClient = apiHttpClient;
+        this.projectKey = projectKey;
+        this.associateId = associateId;
+        this.businessUnitKey = businessUnitKey;
+        this.ID = ID;
+    }
+
+    public ByProjectKeyAsAssociateByAssociateIdInBusinessUnitKeyByBusinessUnitKeyShoppingListsByIDGet get() {
+        return new ByProjectKeyAsAssociateByAssociateIdInBusinessUnitKeyByBusinessUnitKeyShoppingListsByIDGet(
+            apiHttpClient, projectKey, associateId, businessUnitKey, ID);
+    }
+
+    public ByProjectKeyAsAssociateByAssociateIdInBusinessUnitKeyByBusinessUnitKeyShoppingListsByIDHead head() {
+        return new ByProjectKeyAsAssociateByAssociateIdInBusinessUnitKeyByBusinessUnitKeyShoppingListsByIDHead(
+            apiHttpClient, projectKey, associateId, businessUnitKey, ID);
+    }
+
+    public ByProjectKeyAsAssociateByAssociateIdInBusinessUnitKeyByBusinessUnitKeyShoppingListsByIDPost post(
+            com.commercetools.api.models.shopping_list.ShoppingListUpdate shoppingListUpdate) {
+        return new ByProjectKeyAsAssociateByAssociateIdInBusinessUnitKeyByBusinessUnitKeyShoppingListsByIDPost(
+            apiHttpClient, projectKey, associateId, businessUnitKey, ID, shoppingListUpdate);
+    }
+
+    public ByProjectKeyAsAssociateByAssociateIdInBusinessUnitKeyByBusinessUnitKeyShoppingListsByIDPostString post(
+            final String shoppingListUpdate) {
+        return new ByProjectKeyAsAssociateByAssociateIdInBusinessUnitKeyByBusinessUnitKeyShoppingListsByIDPostString(
+            apiHttpClient, projectKey, associateId, businessUnitKey, ID, shoppingListUpdate);
+    }
+
+    public ByProjectKeyAsAssociateByAssociateIdInBusinessUnitKeyByBusinessUnitKeyShoppingListsByIDPost post(
+            UnaryOperator<com.commercetools.api.models.shopping_list.ShoppingListUpdateBuilder> op) {
+        return post(op.apply(com.commercetools.api.models.shopping_list.ShoppingListUpdateBuilder.of()).build());
+    }
+
+    public ByProjectKeyAsAssociateByAssociateIdInBusinessUnitKeyByBusinessUnitKeyShoppingListsByIDDelete delete() {
+        return new ByProjectKeyAsAssociateByAssociateIdInBusinessUnitKeyByBusinessUnitKeyShoppingListsByIDDelete(
+            apiHttpClient, projectKey, associateId, businessUnitKey, ID);
+    }
+
+    public <TValue> ByProjectKeyAsAssociateByAssociateIdInBusinessUnitKeyByBusinessUnitKeyShoppingListsByIDDelete delete(
+            TValue version) {
+        return delete().withVersion(version);
+    }
+
+}